+++ src/main/java/com/thanh/foodOrder/dto/AI/ProductAiResponse.java
@@ -1,8 +1,10 @@
 package com.thanh.foodorder.dto.AI;
 
+import java.math.BigDecimal;
+
 public record ProductAiResponse(
-        Long id,
-        String name,
-        Double price,
-        String img) {
+                Long id,
+                String name,
+                BigDecimal price,
+                String img) {
 }
